+++ dts/digest/index.d.ts
@@ -1,0 +1,2 @@
+export * from "./digest";
+export * from "./full-dump";